Manual black_list better than automatic
Play around.
Fix synthesiser (now black programs are not even added to the queue).
See that a manual black_list is better than an automatic, because in the
automatic we only have identity functions, whereas in the manual list we
can pack verbose synonyms as well, and the same pattern occurs often
(and they are not very efficient to generate + the generated examples
are not as good as naive examples).